fix: move 7702 out of shared
