+++ examples/ui-demo/src/components/7702/MintCard7702.tsx
