+++ examples/ui-demo/src/components/7702/MintStages.tsx
